pyproject.toml
[project]
name = "zotero-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=0.3.0",
    "python-dotenv>=1.0.0",
    "pyzotero>=1.6.11",
]
